pythonMailing/requirements.txt
feedparser
python-dotenv
schedule
beautifulsoup4
googletrans==4.0.0-rc1
requests
newspaper3k
googlenewsdecoder
lxml_html_clean
